length of list

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -8,7 +8,7 @@
  */
 size_t list_len(const list_t *h)
 {
-	unsigned int count = 1;
+	size_t count = 1;
 
 	if (h->next == NULL)
 		return (1);
